// frontend/src/api-generated/models/viTriRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface ViTriRequest {
  maViTri?: string;
  tenViTri: string;
  tenTiengAnh?: string;
  loaiViTri?: string;
  sucChuaToiDa?: number;
  dienTichM2?: number;
  chieuCaoM?: number;
  capDoBaoMat?: string;
  laPhongKinh?: boolean;
  coDieuHoaTrungTam?: boolean;
  coHeThongPccc?: boolean;
  coKiemSoatCua?: boolean;
  moTaChiTiet?: string;
  trangThai?: string;
}